requirements.txt
flask==2.3.6
gunicorn==20.1.0
celery==5.3.1
redis==4.5.5
sqlalchemy==1.4.49
psycopg2-binary==2.9.7
pdfminer.six==20221105
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0
 
